requirements.txt
alive_progress==2.1.0
mysql-connector-python==8.3.0
openpyxl==3.1.2
pandas==1.1.5
PyYAML==6.0.1
watchdog==2.3.1